hngstage2/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
requests==2.31.0
python-decouple==3.8
Pillow==10.4.0
mysqlclient==2.2.0
